source_code/backend/requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
boto3==1.24.52
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djongo==1.3.6
drf-yasg==1.20.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pycparser==2.21
pymongo==3.12.3
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.9
websocket==0.2.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
pandas==2.1.1
pdfkit==1.0.0
moviepy==1.0.3
django-sslserver==0.22